requirements.txt
streamlit
st-annotated-text
pandas
numpy
ydata-profiling
streamlit-pandas-profiling
Jinja2
